+++ src/bssunfold/core/unfold_combined.py
@@ -64,6 +64,7 @@ def unfold_combined(
     from .unfold_mlem import unfold_mlem
     from .unfold_mlem_odl import unfold_mlem_odl
     from .unfold_qpsolvers import unfold_qpsolvers
+    from .unfold_mystic import unfold_mystic
     from .unfold_doroshenko import unfold_doroshenko
     from .unfold_kaczmarz import unfold_kaczmarz
     from .unfold_lmfit import unfold_lmfit
@@ -96,6 +97,7 @@ def unfold_combined(
             'mlem': unfold_mlem,
             'mlem_odl': unfold_mlem_odl,
             'qpsolvers': unfold_qpsolvers,
+            'mystic': unfold_mystic,
             'doroshenko': unfold_doroshenko,
             'kaczmarz': unfold_kaczmarz,
             'lmfit': unfold_lmfit,